+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # libsunspec changelog
 
 ## v0.1.5 - 2024-11-03
-* added function to enable/disable link error recovery
+* added function to enable/disable error recovery
 * removed read/write register loop
 
 ## v0.1.4 - 2024-11-02